backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
firebase-admin>=6.2.0
django-cors-headers>=4.3.1
Pillow>=10.0.0
requests>=2.31.0
google-api-python-client>=2.126.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
openai>=1.12.0
whitenoise>=6.6.0
gunicorn>=21.2.0
thirdweb-sdk>=3.0.0,<4.0.0
eth-account>=0.5.7,<0.6.0
cloudinary>=1.39.0 